backend/requirements/base.txt
# Django
Django==5.1.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2

# 数据库
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# 认证
djangorestframework-simplejwt==5.3.1

# 任务队列（音频处理）
celery==5.3.6
redis==5.0.3

# 音频处理
pydub==0.25.1

# 评论嵌套
django-mptt==0.16.0

# 图片处理
Pillow==10.2.0

# 工具
python-slugify==8.0.4
python-decouple==3.8

# API 文档
drf-yasg==1.21.7
